uVision - tools/uvision.py implementation
