+++ project_generator/tool.py
@@ -20,6 +20,7 @@ from .targets import Targets
 from .builders.gccarm import MakefileGccArmBuilder
 from .builders.uvision import UvisionBuilder
 from .tools.iar import IAREmbeddedWorkbench
+from .tools.uvision import Uvision
 from .exporters.coide import CoideExporter
 from .exporters.gccarm import MakefileGccArmExporter
 from .exporters.uvision import UvisionExporter
@@ -44,9 +45,9 @@ class ToolsSupported:
         'uvision': {
             'toolchain': 'uvision',
             'toolnames': ['uvision'],
-            'exporter': UvisionExporter,
-            'builder': UvisionBuilder,
-            'flasher': UvisionBuilder,
+            'exporter': Uvision,
+            'builder': Uvision,
+            'flasher': Uvision,
         },
         'coide': {
             'toolchain': 'gcc_arm',
